[generation] more bug fixes + consistency tests

+++ sarangi/ext/_short_path.h
@@ -2,6 +2,6 @@
 #define SHORT_PATH_H
 #include <stddef.h>
 
-void dijkstra_impl(size_t start, size_t stop, size_t T, size_t n, const float * restrict x, float * restrict dist, char * restrict visited, int * restrict pred, float param);
+void dijkstra_impl(size_t start, size_t stop, size_t T, size_t n, const float * restrict x, double * restrict dist, char * restrict visited, int * restrict pred, float param);
 
 #endif
